[proliferate] move automation cloud executor under worker
